backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg[binary]
alembic
python-jose[cryptography]
bcrypt>=4.1.0
email-validator
